requirements.txt
absl-py==0.7.1
astor==0.7.1
astunparse==1.6.3
bleach==1.5.0
cachetools==4.1.1
captcha==0.3
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
easydict==1.8
Flask==1.0.2
gast==0.3.3
gevent==20.6.2
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.19.0
h5py==2.10.0
html5lib==0.9999999
idna==2.7
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mock==4.0.2
numpy==1.16.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.3.0.36
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.12.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.23
Werkzeug==0.15.2
wincertstore==0.2
wrapt==1.11.2
zope.event==4.4
zope.interface==5.1.0
